Develop frontend

+++ ArtSite.Core/Services/TierService.cs
@@ -12,12 +12,21 @@ public class TierService : ITierService
     private readonly IProfileService _profileService;
     private readonly ITierRepository _tierRepository;
     private readonly IStorageService _storageService;
+    private ISubscriptionService? _subscriptionService = null;
+    private readonly IArtRepository _artRepository;
 
-    public TierService(ITierRepository tierRepository, IProfileService profileService, IStorageService storageService)
+    public TierService(ITierRepository tierRepository, IProfileService profileService, IStorageService storageService, IArtRepository artRepository)
     {
         _tierRepository = tierRepository;
         _profileService = profileService;
         _storageService = storageService;
+        _artRepository = artRepository;
+    }
+
+    public ITierService Apply(ISubscriptionService subscriptionService)
+    {
+        _subscriptionService = subscriptionService;
+        return this;
     }
 
     public async Task<Tier> CreateTier(string userId, string name, string description, int profileId, int price, int? extends)
@@ -37,18 +46,31 @@ public class TierService : ITierService
 
     public async Task DeleteTier(string userId, int id)
     {
+        if (_subscriptionService == null)
+            throw new NotAppliedException("SubscriptionService");
         var profile = await _profileService.GetProfileByUserId(userId);
         var tier = await _tierRepository.GetTier(id);
         if (tier == null)
             throw new TierException.NotFoundTier();
         if (tier.ProfileId != profile.Id)
             throw new TierException.NotOwnerTier();
-        // TODO: Все тиры которые зависят от этого тира тоже удалить (или выдать ошибку)
+        if ((await _tierRepository.GetChildTiers(id)).Count > 0)
+            throw new TierException.HasChildTiers();
         if (tier.Avatar != null)
         {
             var file = await _storageService.Apply(_profileService).GetFile(tier.Avatar.Value);
             await _storageService.DeleteFile(userId, file.Id);
         }
+        var arts = await _artRepository.GetArtsByTierId(id);
+        foreach (var art in arts)
+        {
+            if (art.TierId == id)
+            {
+                art.TierId = null;
+                await _artRepository.UpdateArt(art);
+            }
+        }
+        await _subscriptionService.UnscribeAll(tier.Id);
         await _tierRepository.DeleteTier(tier.Id);
     }
 
@@ -82,7 +104,7 @@ public class TierService : ITierService
 
     public Task<List<Tier>> GetTiers(int artistId)
     {
-        return _tierRepository.GetTiersByArtist(artistId);
+        return _tierRepository.GetTiersByProfile(artistId);
     }
 
     public async Task UpdateAvatar(string userId, int tierId, IFileUploader fileUploader)
